rename attribute-transition-model.js -> attribute-transition-utils.js
