+++ src/core/lib/attribute-transition-utils.js
